requirements.txt
# Data Processing
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Google AI
google-generativeai==0.3.2

# Data Science
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
